alter types to smaller ones, extract rope to a separate function (oscar would choke on local variables); doesn't work yet because of rope

+++ nnet64.h
@@ -12,6 +12,6 @@
 void softmax(REUPtr x, uint16_t size);
 
 // generate
-float* forward(Transformer* transformer, int token, int pos);
+float* forward(Transformer* transformer, uint16_t token, uint16_t pos);
 
 #endif // NNET_H
